+++ src/components/rightSidePanel/info/TabInfo.vue
@@ -1,31 +1,16 @@
 <template>
-  <div class="space-y-2 rounded-lg bg-interface-surface p-3 text-sm">
-    <div class="flex flex-col gap-1">
-      <span class="font-medium text-base-foreground">{{
-        $t('rightSidePanel.nodeType')
-      }}</span>
-      <span class="text-base-foreground-muted">{{ node.type }}</span>
-    </div>
-    <div v-if="node.id !== undefined" class="flex flex-col gap-1">
-      <span class="font-medium text-base-foreground">{{
-        $t('rightSidePanel.nodeId')
-      }}</span>
-      <span class="text-base-foreground-muted">{{ node.id }}</span>
-    </div>
-    <div v-if="nodeDescription" class="flex flex-col gap-1">
-      <span class="font-medium text-base-foreground">{{
-        $t('rightSidePanel.description')
-      }}</span>
-      <p class="text-base-foreground-muted">{{ nodeDescription }}</p>
-    </div>
+  <div v-if="nodeInfo" class="rounded-lg bg-interface-surface p-3">
+    <NodeHelpContent :node="nodeInfo" />
   </div>
 </template>
 
 <script setup lang="ts">
-import { computed } from 'vue'
+import { computed, watch } from 'vue'
 
+import NodeHelpContent from '@/components/node/NodeHelpContent.vue'
 import type { LGraphNode } from '@/lib/litegraph/src/litegraph'
 import { useNodeDefStore } from '@/stores/nodeDefStore'
+import { useNodeHelpStore } from '@/stores/workspace/nodeHelpStore'
 
 const props = defineProps<{
   nodes: LGraphNode[]
@@ -33,13 +18,20 @@ const props = defineProps<{
 const node = computed(() => props.nodes[0])
 
 const nodeDefStore = useNodeDefStore()
+const nodeHelpStore = useNodeHelpStore()
 
 const nodeInfo = computed(() => {
   return nodeDefStore.fromLGraphNode(node.value)
 })
 
-const nodeDescription = computed(() => {
-  // @ts-expect-error - desc property may exist on constructor
-  return node.value.constructor?.desc || nodeInfo.value?.description || null
-})
+// Open node help when the selected node changes
+watch(
+  nodeInfo,
+  (info) => {
+    if (info) {
+      nodeHelpStore.openHelp(info)
+    }
+  },
+  { immediate: true }
+)
 </script>
